src/requirements.txt
pytorch
torchvision
torchaudio
pandas
scikit-learn
tqdm
ipykernel
albumentations